Dev-Hive-main/requirements.txt
flask
flask-cors
openai
pinecone
python-dotenv
requests
streamlit
cryptography
gunicorn